TemplateWriter.Create
